Lazily use Diagnostic in GenericPart
