+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/deduce/DeduceTypes2.java
@@ -2493,46 +2493,6 @@ public class DeduceTypes2 implements GDeduceTypes2 {
 		}
 	}
 
-	static class GenericPart {
-		private final ClassStatement classStatement;
-		private final TypeName       genericTypeName;
-
-		@Contract(pure = true)
-		public GenericPart(final ClassStatement aClassStatement, final TypeName aGenericTypeName) {
-			classStatement  = aClassStatement;
-			genericTypeName = aGenericTypeName;
-		}
-
-		@Contract(pure = true)
-		public @Nullable TypeNameList getGenericPartFromTypeName() {
-			final NormalTypeName ntn = getGenericTypeName();
-			if (ntn == null)
-				return null;
-			return ntn.getGenericPart();
-		}
-
-		@Contract(pure = true)
-		private @Nullable NormalTypeName getGenericTypeName() {
-			// assert genericTypeName != null;
-			if (genericTypeName == null) {
-
-				tripleo.elijah.util.SimplePrintLoggerToRemoveSoon.println_err_4("1860 who cares // assert genericTypeName != null");
-
-			}
-			/*
-			 * for (boolean aB : _inj().new_boolean[]{genericTypeName != null,
-			 * genericTypeName instanceof NormalTypeName}) { assert aB; }
-			 */
-
-			return (NormalTypeName) genericTypeName;
-		}
-
-		@Contract(pure = true)
-		public boolean hasGenericPart() {
-			return !classStatement.getGenericPart().isEmpty();
-		}
-	}
-
 	static class NullConnector implements IVariableConnector {
 		@Override
 		public void connect(final VariableTableEntry aVte, final String aName) {
